State: Modularize Jetpack Connect actions
