+++ client/state/jetpack-connect/actions/start-authorize-step.js
@@ -1,0 +1,14 @@
+/**
+ * Internal dependencies
+ */
+import { JETPACK_CONNECT_QUERY_SET } from 'state/jetpack-connect/action-types';
+
+import 'state/jetpack-connect/init';
+
+export function startAuthorizeStep( clientId ) {
+	return {
+		type: JETPACK_CONNECT_QUERY_SET,
+		clientId,
+		timestamp: Date.now(),
+	};
+}
